auto translate
added a json containing all kanji and kana in a dict. created a function in the bot_translator cog that detects if there are any kanji or kana in a message, and if so it will auto translate that message and send the result to the channel the original message was sent to. i also did some testing with googletrans lib but it doesnt seem to work for me yet so it is archived for now